rename nats to more implementation agnostic pubsub
